+++ src/Loader.hs
@@ -7,33 +7,28 @@ import Data.Vector (Vector, toList)
 import Data.Map.Ordered (fromList)
 import Data.Csv (decode, HasHeader(..))
 import Data.Text (Text, pack, strip)
-import Data.Text.Read (decimal)
-import Element (Row, Elem(..))
+import Element (toElem)
+import Table (Table(..), Row, empty, buildTable)
 
-loadCsv :: ByteString -> Either Text [Row]
+type Rec = [Text]
+
+loadCsv :: ByteString -> Either Text Table
 loadCsv csvData =
   case decode NoHeader csvData of
     Left err -> Left $ pack err
     Right recs -> Right $ extract $ deepToList recs
 
-extract :: [[Text]] -> [Row]
-extract [] = []
-extract (colNames:recs) = extractRows (fmap strip colNames) recs
+extract :: [Rec] -> Table
+extract [] = empty
+extract (header:recs) = extractRows (fmap strip header) recs
 
--- Csv Column Names -> Csv Row values -> Output Rows
-extractRows :: [Text] -> [[Text]] -> [Row]
-extractRows = fmap . extractRow
+extractRows :: Rec -> [Rec] -> Table
+extractRows header vals = buildTable rows
+  where rows = fmap (extractRow header) vals
 
--- Csv Column Names -> Csv Row values -> Output Row
-extractRow :: [Text] -> [Text] -> Row
-extractRow colNames = fromList . zip colNames . fmap (toElem . strip)
+-- Csv Header -> Csv Value -> Output Row
+extractRow :: Rec -> Rec -> Row
+extractRow header = fromList . zip header . fmap (toElem . strip)
 
 deepToList :: Vector (Vector a) -> [[a]]
 deepToList = toList . fmap toList
-
--- Ultimately this should be a proper value parser
-toElem :: Text -> Elem
-toElem s =
-  case decimal s of
-    Left _ -> SElem s
-    Right (i, _) -> IElem i