+++ src/Runner.hs
@@ -1,40 +1,17 @@
-{-# LANGUAGE OverloadedStrings #-}
-
 module Runner
   ( run
   ) where
 
-import Query (Query, Col, selection, condition, evalSel)
-import Element (Row, Elem(..))
-import QueryException (QueryException(..))
-import Data.Map.Ordered as DMO (lookup, fromList)
-import Control.Exception.Base (throw)
+import Query (Query, condition, evalSel)
+import Table (Row, Table, rows, buildTable)
 import Expression (isTrue)
 
-newtype MissingColumn = MissingColumn Col
-
--- Query -> CSV Rows -> Output Row
-run :: Query -> [Row] -> [Row]
-run query = fmap (select query) . filter (whereFilter query)
-
--- Query Select Columns -> CSV Row -> Selected CSV Row
-select :: Query -> Row -> Row
-select query csvRow =
-  evalSel query csvRow
---  case selectElems selections csvRow of
---    Left (MissingColumn c) -> throw $ MissingColumnsException $ "Table was missing column " <> c
---    Right elems -> DMO.fromList $ zip (evalSel query csvRow) elems
---  where selections = selection query
-
-selectElems :: [Col] -> Row -> Either MissingColumn [Elem]
-selectElems selections csvRow = mapM (toElem csvRow) selections
-
--- CSV Row -> Column name -> Element
-toElem :: Row -> Col -> Either MissingColumn Elem
-toElem csvRow c =
-  case DMO.lookup c csvRow of
-    Nothing -> Left $ MissingColumn c
-    Just e -> Right e
+run :: Query -> Table -> Table
+run query tab = buildTable outRows
+  where outRows = (evalSels . filterWhere) inRows
+        evalSels = fmap (evalSel query)
+        filterWhere = filter (whereFilter query)
+        inRows = rows tab
 
 whereFilter :: Query -> Row -> Bool
 whereFilter query row =