+++ test/fixtures/queries/bool_table.sql
@@ -1,0 +1,3 @@
+SELECT name, NOT isCool AS awesomeButNotCool
+FROM test/fixtures/tables/cool
+WHERE isAwesome