+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for csvsql
 
+## v0.0.8
+* Added Bool support in input tables
+* Added headers in output tables
+
 ## v0.0.7
 * SELECT Expressions
 * SELECT AS